+++ src/verifiers/issuerIdentity/dnsTxt/openAttestationDnsTxtForHedera.test.ts
@@ -1,20 +1,9 @@
 import { v2, v3 } from "@tradetrust-tt/tradetrust";
 import { providers } from "ethers";
-import { documentHederaValidWithDocumentStore } from "../../../../test/fixtures/v2/documentGoerliValidWithDocumentStore";
-import { documentGoerliValidWithToken } from "../../../../test/fixtures/v2/documentGoerliValidWithToken";
-import { documentMixedIssuance } from "../../../../test/fixtures/v2/documentMixedIssuance";
+import { documentHederaValidWithDocumentStore } from "../../../../test/fixtures/v2/documentHederaValidWithDocumentStore";
 import { getProvider } from "../../../common/utils";
 import { openAttestationDnsTxtIdentityProof } from "./openAttestationDnsTxt";
 
-import v3DidSignedRaw from "../../../../test/fixtures/v3/did-signed.json";
-import v3DnsDidSignedRaw from "../../../../test/fixtures/v3/dnsdid-signed.json";
-import v3DocumentStoreIssuedRaw from "../../../../test/fixtures/v3/documentStore-issued.json";
-import v3TokenRegistryIssuedRaw from "../../../../test/fixtures/v3/tokenRegistry-issued.json";
-
-const v3DidSigned = v3DidSignedRaw as v3.SignedWrappedDocument;
-const v3DnsDidSigned = v3DnsDidSignedRaw as v3.SignedWrappedDocument;
-const v3DocumentStoreIssued = v3DocumentStoreIssuedRaw as v3.WrappedDocument;
-const v3TokenRegistryIssued = v3TokenRegistryIssuedRaw as v3.WrappedDocument;
 
 class CustomJsonRpcProvider extends providers.JsonRpcProvider {
   async detectNetwork() {
